Add main.cpp modifications
* Add coefficientes for the proportional, integral and derivative
errors. These coeffs were tuned manually.
* If the integral coefficient is greater than cero (either a big or
small value) the car will drive in circles and it goes out the
track very fast; so it should be zero
* If the proportional coeff is around 1.0, the car can go easier
out of the track, so 0.15 is a good value for that.
* When the derivative coeff is around 3 or 4; it produces large
steering angles and the driving turn really rough. By reducing it,
the car drives in a smoother and more human mode.
* Calculate the steer value depending on the CTE value
* Check for valid values of steering [-1, 1]